Add windows back
Several tests are flaky and the performace of the windows agents meant
they where more likely than not to flake.

However I am observing these flakes in a different CI system running
Linux - so they are there and they are not windows specific.  so it is
better to have the report that the code is bad than to completely ignore
it.